[#2] New Focus Button in Header
New focus button now brings up a bootstrap modal with our new focus form. The bootstrap modal is a template that we load right away (it needs to be in the DOM I think).
Updated CSS for modal. (The overflow:hidden on the body made things jerk b/c it got rid of the scrollbar).
Moved submit form event handler code from new_focus in focus_list to the modal.