[#4] User can edit / delete focus
Meteorized the focusModal. Added a template helper 'focus' that gets 'currentFocus' out of Session. I set that when the user hits 'edit' for a focus. That data is reactive, so whenever it gets set the modal updates!
When we hide the modal, I Session.set('currentFocus', null), and removeData().
The removeData() fix solved an issue where data-action would be set to 'New' and there would be no data-focus-id, but you could still grab them. This meant after hitting edit, you couldn't create a new focus, you'd just update the last one. Clearing out data seemed like the easiest thing.
I'd rather have access to the focus directly in the 'submit form' event handler, but that doesn't seem to work. I just searched on this and if you use #with, the data context is set. The problem is there is no data context when modal is first loaded, it is undefined! So even though we add focus after the fact, we can't access it in the event handler.