[#4] Provided data context to focusModal
So I wanted the focusModal to have a context so I could reference stuff via 'this' in the event handlers. i.e. do we have a focus? Check this._id. This didn't work when I used 'focus' template helper.
BEST SOLUTION: We pass in data context to focusModal when we first render it. I created a template helper in focuslist.js for currentFocus that will update as currentFocus updates.