[#15] USer can add tasks by pressing enter key
Abstracted out the click .new-task event code to addNewTask function so we don't have to write it twice.
Then I added a new event handler 'keypress input[new-task-title]' that checks if enter was pressed, then calls addNewTask.
I went with addNewTask.call(this, e), b/c that seemed like the most appropriate. We pass the current this context and the event.