[#14] User can complete tasks
Added taskItem template helpers and event handlers. The helpers check if task is completed so we can check the box and change the color.
The event handler is on the checkbox and updates the task status. Eventually, we will exclude completed tickets in general, but for now I just grey it out and check it.